Moved two functions to `test_helper.py`
